Refactor events/import_entity_events
